Move entire post_connect to base ChargePoint
